requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.2

pydantic-settings==2.4.0
python-multipart==0.0.9
bcrypt==4.1.2
PyJWT==2.8.0

httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
asgi-lifespan==2.1.0
